LPS-180879 Extract ScreenNavigationEntry methods from CountryRegionsScreenNavigationCategory to CountryRegionsScreenNavigationEntry, extends CountryRegionsScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
